Moved the dupicated code from ScannerWithFlashlightActivity to QRScannerFragment
